+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed sources to correctly handle sample rate and channel count changes at span boundaries.
 - Fixed sources to detect parameter updates after mid-span seeks.
 - Fixed `Stoppable` and `Skippable` not signaling exhaustion.
+- Fixed `Speed::try_seek` to pass seek position directly to inner source without accounting for speed factor (#876).
 
 ## Version [0.22.2] (2026-02-22)